feat: isolates LanguageLinter component
