+++ src/components/Components.css
@@ -1,0 +1,223 @@
+* {
+  box-sizing: border-box;
+}
+
+a {
+  color: #0064c9;
+}
+
+.suggestion-list {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  margin-top: 20px;
+}
+
+.suggestion-container {
+  display: flex;
+  gap: 10px;
+  flex-direction: column;
+  padding: 16px 16px 10px;
+  border: 1px solid #e0e0e0;
+  background: #ffffff;
+  box-shadow: 0px 2px 0px rgba(0, 0, 0, 0.03);
+  border-radius: 6px;
+  transition: 0.3s all ease-out;
+}
+
+.suggestion-container.suggestion-dismissed {
+  display: none;
+}
+
+.suggestion-container.suggestion-accepted {
+  display: none;
+}
+
+.suggestion-rule-container {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+}
+
+.suggestion-rule-severity {
+  display: inline-block;
+  width: 6px;
+  height: 6px;
+  background-color: #919797;
+  border-radius: 100%;
+}
+
+.suggestion-rule-severity.severe {
+  background-color: #df2d24;
+}
+
+.suggestion-rule-severity.moderate {
+  background-color: #0c74df;
+}
+
+.suggestion-rule-label {
+  font-weight: normal;
+  font-size: 10px;
+  line-height: 135%;
+  letter-spacing: 0.75px;
+  text-transform: uppercase;
+  color: #6e7575;
+}
+
+.suggestion-summary-container {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+}
+
+.summary-basic {
+  font-size: 15px;
+  color: #283636;
+  font-weight: 700;
+  gap: 6px;
+}
+
+.basic-summary-example {
+  font-weight: 400;
+}
+
+.suggestion-summary-current {
+  text-decoration: line-through;
+  color: #df2d24;
+}
+
+.suggestion-summary-current-text {
+  font-size: 15px;
+  color: #283636;
+}
+
+.suggestion-summary-arrow {
+  transform: scale(0.9);
+}
+
+.summary-removal .suggestion-summary-current {
+  padding: 5px 7px 6px;
+  color: #fff;
+  background-color: #df2d24;
+  text-shadow: 0 1px rgb(0 0 0 / 10%);
+  border-radius: 6px;
+  transition: all 0.065s ease-out;
+}
+
+.summary-removal .suggestion-summary-current:hover {
+  cursor: pointer;
+  transform: translateY(-1px);
+  background-color: #ec5148;
+}
+
+.summary-removal .suggestion-summary-current:active {
+  transform: translateY(1px);
+}
+
+.summary-removal .suggestion-summary-current-text {
+  font-size: 15px;
+  color: #fff;
+}
+
+.suggestion-summary-replacement {
+  padding: 5px 7px 6px;
+  font-size: 15px;
+  background-color: #3387dd;
+  border-radius: 6px;
+  color: #fff;
+  transition: all 0.065s ease-out;
+}
+
+.suggestion-summary-replacement:hover {
+  cursor: pointer;
+  transform: translateY(-1px);
+  background-color: #4c9beb;
+}
+
+.suggestion-summary-replacement:active {
+  transform: translateY(1px);
+}
+
+.suggestion-summary-replacements {
+  display: flex;
+  flex-direction: row;
+  gap: 6px;
+}
+
+.suggestion-summary-replacements.no-replacement {
+  padding: 5px 7px 6px;
+  font-size: 15px;
+  box-shadow: inset 0 0 0 1px rgba(0 0 0 / 15%);
+  border-radius: 6px;
+  color: rgba(0, 0, 0, 0.5);
+}
+
+.suggestion-description-text {
+  margin: 2px 0 4px;
+  font-size: 13px;
+  line-height: 140%;
+  color: #283636;
+}
+
+.suggestions-ctas {
+  display: flex;
+  justify-content: space-between;
+  margin: -4px -8px 0;
+}
+
+.suggestion-cta-secondary {
+  display: flex;
+  position: relative;
+  padding: 6px 10px;
+  gap: 6px;
+  align-items: center;
+  border: none;
+  background-color: transparent;
+  font-size: 12px;
+  color: #6e7575;
+  border-radius: 3px;
+  transition: all 0.1s ease-out;
+  text-decoration: none;
+}
+
+.suggestion-cta-secondary:hover {
+  background-color: rgba(0, 0, 0, 0.05);
+  cursor: pointer;
+}
+
+.suggestion-cta-secondary.delete-is-hovered:before {
+  content: "Dismiss";
+  display: inline-block;
+  position: absolute;
+  top: -30px;
+  left: 50%;
+  transform: translateX(-50%);
+  padding: 4px 8px;
+  color: #fff;
+  border-radius: 3px;
+  background-color: rgba(0 0 0 / 80%);
+}
+
+.suggestion-cta-secondary.delete-is-hovered:after {
+  content: "";
+  width: 0;
+  height: 0;
+  position: absolute;
+  top: -8px;
+  left: 50%;
+  transform: translateX(-50%);
+  border-top: 3px solid rgba(0 0 0 / 80%);
+  border-right: 3px solid transparent;
+  border-left: 3px solid transparent;
+}
+
+.suggestions-empty-state {
+  display: grid;
+  place-items: center;
+  margin-top: 20px;
+  height: 100%;
+  width: 100%;
+  color: rgba(0, 0, 0, 0.76);
+  border: 2px solid rgba(0, 0, 0, 0.05);
+  border-radius: 8px;
+}
