+++ src/components/personalDictionary.js
